### 2024-05-09 #6740 @editorialbot
[PRE REVIEW]: EPyT-C: A Python package for water quality modeling in water distribution systems
**Submitting author:** @SWIL-IITK (<a href="http://orcid.org/0000-0002-7390-7848">Abhijith G R</a>)
**Repository:** https://github.com/SWIL-IITK/EPyT-C
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
