### 2024-05-09 #6740 @kthyng
- [ ] Clear research application/in scope

In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the database listed above or your own (non-conflicted) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.

I note that this is a resubmission of this previous submission: https://github.com/openjournals/joss-reviews/issues/6348, thanks.
