### 2024-05-28 #6740 @kthyng
Hi @SWIL-IITK
* Installation instructions are given for local installation but I found it on PyPI: https://pypi.org/project/epytc/ — do you plan to keep this up to date?
* We have a backlog of submissions so I will add this to our waitlist. Thanks for your patience.
