### 2024-06-13 #6885 @editorialbot
[REVIEW]: EPyT-C: A Python package for water quality modeling in water distribution systems
**Submitting author:** @SWIL-IITK (<a href="http://orcid.org/0000-0002-7390-7848">Abhijith G R</a>)
**Repository:** https://github.com/SWIL-IITK/EPyT-C
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
