### 2024-07-09 SWIL-IITK/EPyT-C#7 @meghnathomas
JOSS Review
Hi! I will be conducting my JOSS review in this issue.

This is a cool package, thank you for your hard work on it! My initial comments (many cosmetic but some a little more involved) can be found below:

### Landing page
- [ ] The **Code Instructions** section instructs users to install the package using `pip install epyt-c`, but I think that should be `pip install epytc`?
- [ ] Please cite the paper and Github repository of EPyT, since you are presenting an extension of their package!
- [ ] Please define all acronyms the first time they are used (e.g., MSRT, WDS)
- [ ] I was successfully able to run the **run_sample.py** file and obtained an Excel file but I am unsure what the different sheets in this file show and what the data points (columns, rows) are.  I recommend including a clear and brief explanation (perhaps with a figure) explaining how to interpret the results
- [ ] I recommend adding hyperlinks to the README and default_values.yaml files in the epytc folder when you mention them on the landing page under **Sample python script** so that users can easily access them

### Jupyter Notebook examples

- [ ]  I get the following error after cell 3 when I run any of the jupyter notebook files:

Based on [this link](https://stackoverflow.com/questions/39125532/file-does-not-exist-in-jupyter-notebook) I see that replacing the line  `root = os.path.dirname(os.path.realpath(__file__))` with `root = os.path.abspath('')` might be more appropriate for jupyter notebook files. I use Python 3.9 if that helps. Just curious, do these files run without error for the authors?

- [ ] I tested a few of the jupyter notebook files and was able to run them all successfully but I see this at the end of the final cell (this is for EX2):

What is happening here and what is this exception that occurs? If this is meant to happen and all users can expect to see this, I think it would be great to mention this in the docs and/or the landing page and explain why this happens.

- [ ] I think it would be great if, at least in the first example and in run_sample.py, you could read and plot the results (as a time series, if that is what is being shown in the final sheet, or as a network plot if you are showing results for all pipes during one time step) of the water quality analysis from the Excel file to show users what results they can expect and how to read the Excel file.

- [ ] Acknowledgments/credits are different in the paper and on the landing page
- [ ] Please include citation to the EPyT JOSS paper!
- [ ] Do other EPANET interfacing tools like WNTR and OOPNET offer water quality modeling abilities similar to epyt-c? Why did you choose to extend EPyT instead of those tools? (Just curious!)

### Docs
- [ ] Are you continuing to edit them? I see that the API docs are quite solid, but are you also planning to include information from your landing page and examples?
- [ ] The example jupyter notebook files demonstrate a very comprehensive list of the different inputs a user can configure which is very useful! I think it would be great if the authors mirrored these examples in the docs, numbered the same way, and without the cells that repeat in each jupyter notebook. This is just so potential users have a single, well-documented place where they can find the different ways they can customize their simulations.
