### 2024-07-10 dd1337a
Update os-test.yml
