+++ .github/workflows/os-test.yml
@@ -19,13 +19,14 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
+        version: ["3.10", "3.9", "3.8"]
 
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
-        python-version: "3.10"
+        python-version: ${{ matrix.version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
