### 2024-07-10 ac5b8c9
Update os-test.yml
