+++ .github/workflows/os-test.yml
@@ -19,7 +19,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        version: ["3.10", "3.9", "3.8"]
+        version: ["3.11", "3.10", "3.9"]
 
     steps:
     - uses: actions/checkout@v4
