+++ .github/workflows/run_tests.yml
@@ -13,16 +13,20 @@ permissions:
   contents: read
 
 jobs:
-  build:
 
-    runs-on: ubuntu-latest
-
+  build_run:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        version: ["3.11", "3.10", "3.9"]
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: "3.10"
+        python-version: ${{ matrix.version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
