### 2024-08-22 #6885 @cheginit
@SWIL-IITK, after reviewing the feedback from our reviewers and conducting my own analysis, I've identified several key issues that need to be addressed before we can proceed with the review:

- Scholarly effort: A reviewer has pointed out to me via email that the submission "ineffectively reproduces work that has already been long understood in the field and presents cases that are already well-documented in existing software." This raises concerns about the contribution of your work. In JOSS [documents](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) under the "Substantial Scholarly Effort" section, we state that:

> JOSS publishes articles about software that represent substantial scholarly effort on the part of the authors. Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

- Unaddressed Issues: The issue opened by @meghnathomas on July 9 (https://github.com/SWIL-IITK/EPyT-C/issues/7) has not received any response. This issue highlights problems including example notebooks not running as expected.

- Demonstration of Contributions: Considering the first issue, more comprehensive example notebooks are needed to clearly demonstrate the significant contributions of your submission.

To move forward with your submission, please provide the following:

- [ ] A plan detailing how you will address the reviewer's feedback regarding your contributions.
- [ ] A response to @meghnathomas's issue, including a timeline for resolving the reported problems.
- [ ] A proposal for additional example notebooks that will effectively showcase the contributions of your software package.
- [ ] A tentative timeline for implementing these improvements.

Please submit your plan and timeline as a reply to this comment within the next two weeks. This will help us assess the viability of your submission moving forward.
We look forward to your response and are committed to working with you to improve your submission to meet JOSS standards.
