### 2024-09-10 #6885 @kthyng
@SWIL-IITK Ah, a miscommunication. Considering we hadn't heard from you on the [reviewer issue](https://github.com/SWIL-IITK/EPyT-C/issues/7) several months ago nor the message above, we figured you had abandoned this submission. In lieu of restarting this submission, I propose that you address the issues in detail from @cheginit's [comment](https://github.com/openjournals/joss-reviews/issues/6885#issuecomment-2305432376), then resubmit after you've taken all the time necessary to do so. We recognize that this will probably take a substantial amount of time and there is no hurry to resubmit. Thank you.
